requirements.txt
apache-airflow[postgres]
pandas
numpy
praw
awscli
s3fs